Add tests for equippable parts
