+++ test/fixtures/equippableSlotsFixture.ts
